requirements.txt
kivy>=2.1.0
opencv-python>=4.5.0
numpy>=1.19.0
python-chess>=1.0.0
Pillow>=8.0.0
requests>=2.25.0
svglib>=1.5.1
reportlab>=3.6.12
plyer>=2.0.0 